+++ ydb/tests/functional/audit/conftest.py
@@ -2,4 +2,4 @@
 # but somehow it does not
 #
 # for ydb_{cluster, database, ...} fixture family
-pytest_plugins = 'ydb.tests.library.harness.ydb_fixtures'
+pytest_plugins = 'ydb.tests.library.fixtures'
