+++ ydb/tests/functional/serverless/conftest.py
@@ -9,7 +9,7 @@ import contextlib
 # but somehow it does not
 #
 # for ydb_{cluster, database, ...} fixture family
-pytest_plugins = 'ydb.tests.library.harness.ydb_fixtures'
+pytest_plugins = 'ydb.tests.library.fixtures'
 
 
 logger = logging.getLogger(__name__)
