+++ ydb/tests/functional/serverless/ya.make
@@ -22,6 +22,7 @@ DEPENDS(
 PEERDIR(
     contrib/python/tornado/tornado-4
     ydb/tests/library
+    ydb/tests/library/fixtures
     ydb/tests/oss/ydb_sdk_import
     ydb/public/sdk/python
 )
