+++ ydb/tests/functional/tenants/conftest.py
@@ -7,7 +7,7 @@ from ydb.tests.library.clients.kikimr_http_client import HiveClient
 # but somehow it does not
 #
 # for ydb_{cluster, database, ...} fixture family
-pytest_plugins = 'ydb.tests.library.harness.ydb_fixtures'
+pytest_plugins = 'ydb.tests.library.fixtures'
 
 
 @pytest.fixture(scope='module')
