+++ ydb/tests/functional/tenants/ya.make
@@ -25,6 +25,7 @@ DEPENDS(
 PEERDIR(
     contrib/python/requests
     ydb/tests/library
+    ydb/tests/library/fixtures
     ydb/tests/library/clients
     ydb/tests/oss/ydb_sdk_import
     ydb/public/sdk/python
