+++ ydb/tests/library/fixtures/ya.make
@@ -1,0 +1,12 @@
+PY3_LIBRARY()
+
+PY_SRCS(
+    __init__.py
+)
+
+PEERDIR(
+    ydb/tests/library
+    ydb/public/sdk/python
+)
+
+END()
